Moved CSRF utils to the utilities folder for better reusability.
